Added options -T to force taxid of the sequences, and -v for verbose output
When -T is set, for each observed k-mer the taxid of the sequence is set - instead of the
lowest common ancestor of of the sequence taxid and the currently set taxid. This is useful
for setting the taxid of contaminant sequences, which may also be observed in database
genomes, to the contaminant taxid.

-v gives more verbose output