Fix jellyfish detection
On some versions of bash, `command -v jellyfish` outputs the path to jellyfish to stdout and results in $JELLYFISH_BIN having two paths to `jellyfish` separated by a newline (which obviously breaks the `build_db` script). I've modified it here to redirect the stdout along with the stderr to /dev/null which should fix this edge case.