Add functionality to process multiple samples in a row
Add functionality to load a database catted into a single file instead of from multiple files from
a directory.

This combination allows the workflow of loading the database from a compressed file or network pipe
and processing multiple files while maintaining performance because the file is not implicitly
mmap'd and cached implicitly.